Fixed ReadElement() using incorrect elementSize

+++ src/ObjLoading/XModel/Gltf/Internal/GltfAccessor.cpp
@@ -425,7 +425,7 @@ bool UnsignedShortAccessor::GetUnsignedVec4(const size_t index, unsigned (&out)[
         return false;
 
     uint16_t temp[4];
-    if (!m_buffer_view->ReadElement(temp, index, sizeof(uint8_t[4]), m_byte_offset))
+    if (!m_buffer_view->ReadElement(temp, index, sizeof(uint16_t[4]), m_byte_offset))
         return false;
 
     out[0] = static_cast<unsigned>(temp[0]);
